+++ astropy/utils/compat/argparse.py
@@ -1,6 +1,10 @@
 from __future__ import absolute_import
 
-try:
+import sys
+
+if sys.version_info[:2] <= (2, 6):
+    from ._argparse import *
+elif sys.version_info[0] == 3 and sys.version_info[:2] <= (3, 1):
+    from ._argparse import *
+else:
     from argparse import *
-except ImportError:
-    from .argparse_py2 import *